Make migrations work in Rails 3
Heroku updated to postgres 9.5 client on the dyno and the `-i` option has been removed from `pg_dump`. Rails 3 passes that option as part of its `db:structure:dump` task (which runs after migrations), causing the rake task to exit with a non-0 status.

This just no-ops the structure dumping altogether on the dyno.